+++ packages/tux/src/components/TextField/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TextField'
