--- packages/tux/src/components/TuxModal/styles.ts
@@ -1,14 +1,0 @@
-export const InputStyles = {
-    labelTextColor: '#8091A5',
-    borderColor: '#CBCBCB',
-}
-
-export const buttonStyles = {
-    backgroundColor: '#E5EBED',
-    borderColor: '#C3CFD5',
-    textColor: '#536171',
-    greenTheme: {
-        backgroundColor: '#3BB172',
-        borderColor: '#188E18',
-    }
-}
