+++ packages/tux/src/styles.ts
@@ -1,0 +1,30 @@
+export const tuxColors = {
+  textLight: '#FFFFFF',
+  textDark: '#313131',
+  textGray: '#3C4858',
+  colorBlack: '#000000',
+  colorPink: '#E8008A',
+  colorSnow: '#E5EBED',
+  colorPurple: '#473BB1',
+}
+
+export const InputStyles = {
+  labelTextColor: '#8091A5',
+  borderColor: '#CBCBCB',
+}
+
+export const buttonStyles = {
+  backgroundColor: '#E5EBED',
+  borderColor: '#C3CFD5',
+  textColor: '#536171',
+  greenTheme: {
+      backgroundColor: '#3BB172',
+      borderColor: '#188E18',
+  }
+}
+
+export default {
+  tuxColors,
+  InputStyles,
+  buttonStyles
+}
