--- packages/tux/src/styles.tsx
@@ -1,9 +1,0 @@
-export const tuxColors = {
-  textLight: '#FFFFFF',
-  textDark: '#313131',
-  textGray: '#3C4858',
-  colorBlack: '#000000',
-  colorPink: '#E8008A',
-  colorSnow: '#E5EBED',
-  colorPurple: '#473BB1',
-}
